Add helper for token transformations
